+++ src/test/java/us/ihmc/euclid/tuple4D/QuaternionTest.java
@@ -1,6 +1,7 @@
 package us.ihmc.euclid.tuple4D;
 
-import static us.ihmc.robotics.Assert.*;
+import static org.junit.jupiter.api.Assertions.*;
+import static us.ihmc.euclid.tools.EuclidJUnitTools.*;
 
 import static us.ihmc.euclid.testSuite.EuclidTestSuite.*;
 
@@ -134,7 +135,7 @@ public class QuaternionTest extends QuaternionBasicsTest<Quaternion>
          }
          q.setUnsafe(qx, qy, qz, qs);
          newHashCode = q.hashCode();
-         assertNotEquals(newHashCode, previousHashCode);
+         assertNotEquals((long) newHashCode, (long) previousHashCode);
          previousHashCode = newHashCode;
       }
    }
